repl_rt -> altered register_remote_locator [added new policy/ strategy] core_cluster_mgr -> new functionality provides the ability to have more than one primary address to connect too (case sink nodes > source nodes) repl2_rtsource_conn -> innit function that calls the repl2_connection_mgr:connect/2 is now connect/3 with new strategy core_connection_mgr -> state altered, the ep (endpoint) record now has the 'primary' boolean to state whether or not the address is a primary connection core_connection_mgr -> update_endpoints, and filter_blacklisted_endpoints have been modified to deal with the new return type of the Locator (altered at the top) core_connection_mgr -> connection_helper has been modified to continue connecting if the next address in the list is also a primary address (assumes list is ordered primary first secondary last [orddict endpoints])
To be completed:
New supervisor that will take the majoirty of the responsability of rtsource_conn, rtsource_conn will now be a child to that supervisor.
rtsource_conn will only take in a socket, spawn a helper to pull from the RTQ, send over to the sink and receive acknowledgements.
riak_core_connection (gen_fsm) will now have to point to this supervisor rather than rtsource_conn.

RTQ will now need to be altered to have a list of helperPids that are associated with each remote cluster name.
(Other option) RTQ funnels the pushes to the supervisor, and the async_pulls from the helpers are made to the supervisor which in turn sends them too the RTQ.
Finally this fix has broken the rebalancing functionality in rtsource_conn (this functionality will be moved to the supervisor);
Never the less this will have to be fixed. The problem is that it still expects only one connection to be present per sink cluster. This is now not always the case.