Remove extra item in the QEntry tuple for the rtq which will save memory and mean that rules are always followed correctly.
The queue entries previously had their filtering rules stored against them - on a per object basis, eg:
{Seq, NumObjects, Bin, Meta, {BucketName, Remotes}}

Which was wasteful, and not needed as we have an up to date list of the filtering config in the rtq state which changes
when the ring changes.

The end item has been removed and rules will now be checked against state, this has the added benefit in that if we remove
or update a filter rule whilst there are already items in the queue, we will use the updated config from state instead of
relying on outdated config against the object.